pyproject.toml
[project]
name = "block-seek"
version = "0.1.0"
dependencies = [
   "langchain>=0.1.0",
   "openai>=1.12.0",
   "fastapi>=0.110.0",
   "uvicorn>=0.27.0",
   "web3>=6.15.1",
   "eth-account>=0.11.0",
   "eth-utils>=3.0.0",
   "faiss-cpu>=1.7.4",
   "llama-index>=0.10.0",
   "pandas>=2.2.0",
   "numpy>=1.24.0",
   "plotly>=5.19.0",
   "requests>=2.31.0",
   "aiohttp>=3.9.0",
   "python-dotenv>=1.0.0",
   "pydantic>=2.6.0",
   "langchain-openai>=0.3.2",
   "pydantic-settings>=2.7.1",
   "langchain-community>=0.3.15",
]


[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.2.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/block-seek"]
